backend-python/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
openai>=1.10.0
pinecone>=3.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
langdetect>=1.0.9
